tests/ci: SSRF routing test, migration-drift guard, blocking pip-audit
- Add tests/test_fetcher_ssrf.py: feed discovery routes through the SSRF-
  guarded safe_fetch and is_safe_url fails closed on unresolvable hosts.
- Add tests/test_migrations_match_models.py: run alembic upgrade head into a
  throwaway DB and assert models match migrations (catches create_all-hidden
  drift). Skips without CREATEDB.
- Register the 'migration' pytest marker.
- config/settings.py: ignore unknown env vars so a local .env (e.g.
  POSTGRES_PASSWORD) no longer crashes pytest import.
- ci.yml: add a frontend job (typecheck + test + build); make pip-audit
  blocking (no known vulnerabilities in the pinned tree).

+++ config/settings.py
@@ -21,7 +21,7 @@ class Settings(BaseSettings):
 
     # NER service tuning (the separate `ner` worker)
     ner_batch_size: int = 50
-    ner_poll_interval: float = 2.0
+    ner_poll_interval: int = 2
     ner_zombie_min: int = 5
     ner_max_retries: int = 3
